Added a React Router to our frontend
The router is used to navigate through the pages / components. For
this some components had to be changed. Structured the src files:
pages can be found under /src/components/pages/{page_name}
Components that are only used by one page, are stored in the page
directory. Also added a new page: PersonsInVehicle

Started: #56