Configure PSQL for Django
The PSQL database allows us to store information in a more permanent
way, as well as emulate access to the customer's database. It is
configured through the backend/models.py file, andaccessed through
settings.py, which also serves as a proxy to access thedatabase
password contained in secrets.json (not commited for obvious
reasons).