Added a fully working feedback when exceeding the limit.
This enables a feedback to the user when trying to add more items that exceeds the legal limit. It also disables the checkout button if the cart is empty or the limit is exceeded. Still need to be fixed is the deleting of an item. This does currently not update the state.